squash: move new pageobjects into pageobjects dir
