+++ src/test/java/org/jitsi/meet/test/InviteTest.java
@@ -3,8 +3,8 @@ package org.jitsi.meet.test;
 import org.jitsi.meet.test.base.AbstractBaseTest;
 import org.jitsi.meet.test.base.Participant;
 import org.jitsi.meet.test.util.MeetUtils;
-import org.jitsi.meet.test.web.DialInNumbersPage;
-import org.jitsi.meet.test.web.InfoDialog;
+import org.jitsi.meet.test.pageobjects.web.DialInNumbersPage;
+import org.jitsi.meet.test.pageobjects.web.InfoDialog;
 import org.jitsi.meet.test.web.WebParticipant;
 import org.openqa.selenium.WebDriver;
 import org.testng.SkipException;
