+++ src/test/java/org/jitsi/meet/test/pageobjects/web/InfoDialog.java
@@ -1,7 +1,8 @@
-package org.jitsi.meet.test.web;
+package org.jitsi.meet.test.pageobjects.web;
 
 import org.jitsi.meet.test.util.MeetUIUtils;
 import org.jitsi.meet.test.util.TestUtils;
+import org.jitsi.meet.test.web.WebParticipant;
 import org.openqa.selenium.By;
 import org.openqa.selenium.Keys;
 import org.openqa.selenium.WebDriver;
