+++ src/test/java/org/jitsi/meet/test/web/WebParticipant.java
@@ -17,6 +17,8 @@ package org.jitsi.meet.test.web;
 
 import org.jitsi.meet.test.base.*;
 import org.jitsi.meet.test.pageobjects.web.ChatPanel;
+import org.jitsi.meet.test.pageobjects.web.DialInNumbersPage;
+import org.jitsi.meet.test.pageobjects.web.InfoDialog;
 import org.jitsi.meet.test.util.*;
 import org.openqa.selenium.*;
 import org.openqa.selenium.interactions.*;
